Removing unused code causing performance problems
The study details query causes performance problems since it selects table scans the whole nd_experiement table. This was mainly due to the use of the hasRows method in the StudyDetails class. This data is actually not needed as it is not used any where in the application.

Removing used code that required this information.

issue: BMS-17776
reviewer: NaymeshM