Handle redirection to login if xAuthToken invalid
Nursery and Trial now need BMSAPI token to use VariableCacheController
invalidation endpoint. Upon entering Nursery or Trial manager check if
xAuthToken is valid and if it's not, redirect to login to get a new token.
This check is also added to global ajaxError handler in case the user is
already browsing a study, although this is an edge case.

Issue: BMS-3535