Instead of passing variable type value as string, passing it as integer.
This way it will match the switch case to re-instantiate ids of traits from sortVariableIdsAndNames method which is called while removing traits. Before that, it was not matching removed TRAIT's cvtermId string value with integer value and so trait's cvtermId was passed as null which result in NPE in SettingsUtil.setSettingDetailRoleAndVariableType.

issue: BMS-3815
reviewer: Sunny