Open ui-select combo down
In measurements (updateExperimentCell) we were using select2 which
places the widget outside the table, and it's not affected by the
overflow. In subObs the widget is inside the cell and then it's hidden
when it's opened upwards. Opening it downwards minimizes the problem
without adding too much complexity

Issue: IBP-2173
Reviewer: None