Refactor navigateToSubObsTab: add "reload" param
Refactor extra params into "options"
Pass optional "reload" param to $state.transitionTo()
Make navigation without reloading the default: If you are already on the
route, do nothing. This way is more explicit that we are forcing a
reload in some places. Rewrite those to use the param.
Use this new option to open the obs tab without reloading. This will
make possible to either execute start the prepare plating process if we
are already there, or to navigate there first.

Issue: IBP-3669